smartrag/requirements.txt
# ── SmartRAG Python Dependencies ─────────────────────────────────────
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Endee Vector Database SDK
endee>=0.1.0

# Embeddings
sentence-transformers>=2.2.0
torch>=2.0.0

# LLM - Google Gemini
google-generativeai>=0.8.0

# Document Parsing
PyPDF2>=3.0.0
python-docx>=1.0.0

# Frontend
streamlit>=1.38.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
